// src/plugins/common/file-cache-plugin/gql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: any; output: any };
  JSON: { input: any; output: any };
  Money: { input: number; output: number };
  Upload: { input: any; output: any };
};
// ...
export type CreateFileCacheInput = {
  id: Scalars['ID']['input'];
  type: FileCacheType;
  value: Scalars['String']['input'];
};
// ...
export type DateOperators = {
  after?: InputMaybe<Scalars['DateTime']['input']>;
  before?: InputMaybe<Scalars['DateTime']['input']>;
  between?: InputMaybe<DateRange>;
  eq?: InputMaybe<Scalars['DateTime']['input']>;
  isNull?: InputMaybe<Scalars['Boolean']['input']>;
};
// ...
export type FileCache = Node & {
  __typename?: 'FileCache';
  createdAt: Scalars['DateTime']['output'];
  id: Scalars['ID']['output'];
  type: FileCacheType;
  updatedAt: Scalars['DateTime']['output'];
  value: Scalars['String']['output'];
};

export type FileCacheFilterParameter = {
  _and?: InputMaybe<Array<FileCacheFilterParameter>>;
  _or?: InputMaybe<Array<FileCacheFilterParameter>>;
  createdAt?: InputMaybe<DateOperators>;
  id?: InputMaybe<IdOperators>;
  type?: InputMaybe<StringOperators>;
  updatedAt?: InputMaybe<DateOperators>;
  value?: InputMaybe<StringOperators>;
};

export type FileCacheList = PaginatedList & {
  __typename?: 'FileCacheList';
  items: Array<FileCache>;
  totalItems: Scalars['Int']['output'];
};

export type FileCacheListOptions = {
  /** Allows the results to be filtered */
  filter?: InputMaybe<FileCacheFilterParameter>;
  /** Specifies whether multiple top-level "filter" fields should be combined with a logical AND or OR operation. Defaults to AND. */
  filterOperator?: InputMaybe<LogicalOperator>;
  /** Skips the first n results, for use in pagination */
  skip?: InputMaybe<Scalars['Int']['input']>;
  /** Specifies which properties to sort the results by */
  sort?: InputMaybe<FileCacheSortParameter>;
  /** Takes n results, for use in pagination */
  take?: InputMaybe<Scalars['Int']['input']>;
};

export type FileCacheSortParameter = {
  createdAt?: InputMaybe<SortOrder>;
  id?: InputMaybe<SortOrder>;
  updatedAt?: InputMaybe<SortOrder>;
  value?: InputMaybe<SortOrder>;
};

export enum FileCacheType {
  JSON = 'JSON',
  TEXT = 'TEXT',
}
// ...
export type IdOperators = {
  eq?: InputMaybe<Scalars['String']['input']>;
  in?: InputMaybe<Array<Scalars['String']['input']>>;
  isNull?: InputMaybe<Scalars['Boolean']['input']>;
  notEq?: InputMaybe<Scalars['String']['input']>;
  notIn?: InputMaybe<Array<Scalars['String']['input']>>;
};
// ...
export enum LogicalOperator {
  AND = 'AND',
  OR = 'OR',
}
// ...
export type Node = {
  id: Scalars['ID']['output'];
};
// ...
export enum SortOrder {
  ASC = 'ASC',
  DESC = 'DESC',
}
// ...
export type StringOperators = {
  contains?: InputMaybe<Scalars['String']['input']>;
  eq?: InputMaybe<Scalars['String']['input']>;
  in?: InputMaybe<Array<Scalars['String']['input']>>;
  isNull?: InputMaybe<Scalars['Boolean']['input']>;
  notContains?: InputMaybe<Scalars['String']['input']>;
  notEq?: InputMaybe<Scalars['String']['input']>;
  notIn?: InputMaybe<Array<Scalars['String']['input']>>;
  regex?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type UpdateFileCacheInput = {
  id: Scalars['ID']['input'];
  value: Scalars['String']['input'];
};